+++ src/main/java/kitchenpos/menu/service/MenuService.java
@@ -4,9 +4,9 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
-import kitchenpos.application.dto.request.MenuProductRequest;
-import kitchenpos.application.dto.request.MenuRequest;
-import kitchenpos.application.dto.response.MenuResponse;
+import kitchenpos.menu.service.dto.MenuProductRequest;
+import kitchenpos.menu.service.dto.MenuRequest;
+import kitchenpos.menu.service.dto.MenuResponse;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menugroup.domain.MenuGroup;
 import kitchenpos.menu.domain.MenuProduct;
