+++ src/main/java/kitchenpos/menu/service/dto/MenuResponse.java
@@ -1,4 +1,4 @@
-package kitchenpos.application.dto.response;
+package kitchenpos.menu.service.dto;
 
 import java.math.BigDecimal;
 import java.util.List;
