+++ src/main/java/kitchenpos/menugroup/presentation/MenuGroupRestController.java
@@ -3,8 +3,8 @@ package kitchenpos.menugroup.presentation;
 import java.net.URI;
 import java.util.List;
 import kitchenpos.menugroup.service.MenuGroupService;
-import kitchenpos.application.dto.request.MenuGroupRequest;
-import kitchenpos.application.dto.response.MenuGroupResponse;
+import kitchenpos.menugroup.service.dto.MenuGroupRequest;
+import kitchenpos.menugroup.service.dto.MenuGroupResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
