+++ src/main/java/kitchenpos/order/presentation/OrderRestController.java
@@ -3,9 +3,9 @@ package kitchenpos.order.presentation;
 import java.net.URI;
 import java.util.List;
 import kitchenpos.order.service.OrderService;
-import kitchenpos.application.dto.request.OrderRequest;
-import kitchenpos.application.dto.request.OrderStatusRequest;
-import kitchenpos.application.dto.response.OrderResponse;
+import kitchenpos.order.service.dto.OrderRequest;
+import kitchenpos.order.service.dto.OrderStatusRequest;
+import kitchenpos.order.service.dto.OrderResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
