+++ src/main/java/kitchenpos/order/service/dto/OrderLineItemRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.application.dto.request;
+package kitchenpos.order.service.dto;
 
 public class OrderLineItemRequest {
