+++ src/main/java/kitchenpos/order/service/dto/OrderStatusRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.application.dto.request;
+package kitchenpos.order.service.dto;
 
 public class OrderStatusRequest {
