+++ src/main/java/kitchenpos/product/presentation/ProductRestController.java
@@ -3,8 +3,8 @@ package kitchenpos.product.presentation;
 import java.net.URI;
 import java.util.List;
 import kitchenpos.product.service.ProductService;
-import kitchenpos.application.dto.request.ProductRequest;
-import kitchenpos.application.dto.response.ProductResponse;
+import kitchenpos.product.service.dto.ProductRequest;
+import kitchenpos.product.service.dto.ProductResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
