+++ src/main/java/kitchenpos/tablegroup/service/TableGroupService.java
@@ -3,9 +3,9 @@ package kitchenpos.tablegroup.service;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.application.dto.request.OrderTableIdRequest;
-import kitchenpos.application.dto.request.TableGroupRequest;
-import kitchenpos.application.dto.response.TableGroupResponse;
+import kitchenpos.tablegroup.service.dto.OrderTableIdRequest;
+import kitchenpos.tablegroup.service.dto.TableGroupRequest;
+import kitchenpos.tablegroup.service.dto.TableGroupResponse;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.tablegroup.domain.TableGroup;
 import kitchenpos.table.domain.OrderTableRepository;
