+++ src/test/java/kitchenpos/application/TableServiceTest.java
@@ -5,13 +5,13 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.assertj.core.api.SoftAssertions.assertSoftly;
 
 import java.util.List;
-import kitchenpos.application.dto.request.OrderStatusRequest;
-import kitchenpos.application.dto.request.TableGroupRequest;
+import kitchenpos.order.service.dto.OrderStatusRequest;
+import kitchenpos.tablegroup.service.dto.TableGroupRequest;
 import kitchenpos.table.service.dto.TableRequest;
-import kitchenpos.application.dto.response.MenuGroupResponse;
-import kitchenpos.application.dto.response.MenuResponse;
-import kitchenpos.application.dto.response.OrderResponse;
-import kitchenpos.application.dto.response.ProductResponse;
+import kitchenpos.menugroup.service.dto.MenuGroupResponse;
+import kitchenpos.menu.service.dto.MenuResponse;
+import kitchenpos.order.service.dto.OrderResponse;
+import kitchenpos.product.service.dto.ProductResponse;
 import kitchenpos.table.service.dto.TableResponse;
 import kitchenpos.menu.service.MenuService;
 import kitchenpos.menugroup.service.MenuGroupService;
