+++ src/test/java/kitchenpos/supports/ProductFixture.java
@@ -1,7 +1,7 @@
 package kitchenpos.supports;
 
 import java.math.BigDecimal;
-import kitchenpos.application.dto.request.ProductRequest;
+import kitchenpos.product.service.dto.ProductRequest;
 import kitchenpos.common.vo.Price;
 
 public class ProductFixture {
